Update gnome-shell to 41.9
Building gnome-shell from version 41 requires schemas from
gsettings-desktop-schemas version 41.0. Slackware 15.0 supplies schemas
from version 40.0, which will not work. This build installs the needed
41.0 schemas in a custom location and uses a wrapper script to make
gnome-shell aware of the schemas at runtime.

The source download and md5sum for gsettings-desktop-schemas-41.0 are
included in the .info file. Compiling and cleanup of the compiled
schemas is handled in the respective doinst.sh and douninst.sh scripts.