GDM build improvements
Added a wrapper script when starting gdm to drop the '-nodaemon' flag
that Slackware passes when it starts gdm in /etc/rc.d/rc.4. This avoids
the annyoing step of having to manually edit rc.4 to delete the flag,
which has the chance of being over-written if the package is reinstalled
or updated. With the wrapper this issue is effectively negated and the
README instructions can be simplified with the warning removed.

Also added .new file handling to be nicer to the /etc/gdm/custom.conf
file, which was getting overwritten. Thanks to the pointer from
LuckyCyborg.