Update README.md instructions.
I removed avahi from these builds a while ago, since it wasn't actually
a dependency for any of the software built in the queue. There are GNOME
programs that depend on avahi (e.g. gnome-user-share), but for now we
dont need to have it in the queue, and we can simplify the instructions
around that.