Add "SBo-builds-needing-upgrades".
There are a number of builds from other maintainers that need upgrads to
remove the libsoup2/webkit2gtk4.0 dependency:

evolution-data-server - upgraded to 3.48.1 to remove libgdata dep, and use
webkit2gtk4.1

geoclue2 - upgraded to 2.7.x branch to drop the libsoup2 dep.

gnome-online-accounts - upgraded to 3.48.1 to match
evolution-data-server, and switched to use rest and webkit2gtk4.1 to
drop libsoup2 dep.

libchamplain - upgraded to 0.12.21 to use libsoup3. Needed as a dep for
evolution.

evolution - needs to be upgraded to drop libsoup3. Currently FTB when
tested with patches.